requirements.txt
cryptography>=41.0.0
pyperclip>=1.8.2
argon2-cffi>=23.1.0
zxcvbn>=4.4.28
